Initial Commit
This repository shall serve as a base for further development of the HMI.
It implements roughly the MVC aspect.
Model: models/
View:  views/
Controller: routes/

There are also some testing functions included.
It does not include node_modules.

As of now, the following node_modules are used:
* async
* express
* jade
* node-opcua

The structure is based on the automatic generated Express-Jade-Stylus
Template from Eclipse.

In using Stylus/Less or something else, the last word has not yet been
spoken.

Best regards
 Thomas Frei